pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pbprompt"
version = "1.0.1"
description = "A program to register and categorize AI prompts to keep tries."
authors = [{ name = "PBSoft" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
keywords = ["prompts", "AI", "PyQt5", "organizer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "PyQt5>=5.15",
    "ruamel.yaml>=0.18",
    "deep-translator>=1.11",
    "platformdirs>=4.2",
    "requests>=2.31",
]

[project.scripts]
pbprompt = "pbprompt.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-qt>=4.2",
    "ruff>=0.4",
    "pre-commit>=3.6",
    "mypy>=1.8",
    "Sphinx>=7.2",
    "sphinx-rtd-theme>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"pbprompt" = ["py.typed"]

# ---------------------------------------------------------------------------
# Ruff – linter & formatter
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "tests"]
extend-exclude = [
    "src/pbprompt/gui/ui_*.py",
    "src/pbprompt/gui/resources_rc.py",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "C4", "PIE", "SIM"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S"]
# Qt API methods have camelCase names imposed by the framework
"src/pbprompt/gui/models.py" = ["N802"]

[tool.ruff.format]
quote-style = "double"

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
files = ["src"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
